requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic[email]
PyPDF2
python-docx
langchain
langchain-groq
python-dotenv
pydantic-settings
